GraphologyAI/Capstone/ai-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.0
opencv-python-headless==4.8.0.76
pillow==10.0.0
